% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-gene-symbols.R
\name{check_gene_symbols}
\alias{check_gene_symbols}
\title{Check gene symbols}
\usage{
check_gene_symbols(genes, species = "human", conversion = NULL)
}
\arguments{
\item{genes}{One or more gene symbols.}

\item{species}{Species of gene symbols}

\item{conversion}{Validation table, used to crosscheck for items in column \code{symbol}.}
}
\value{
A data frame containing the input gene symbols not matching any entry in the conversion table.
}
\description{
Check gene symbols
}
