% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensembl}
\alias{ensembl}
\title{Conversion from Ensembl ID to gene symbol}
\format{
Data frame with columns for conversion/validation using Ensembl ID.
\describe{
\item{ensembl}{Ensembl ID}
\item{symbol}{Gene symbol}
}
\usage{
ensembl
}
\description{
Conversion from Ensembl ID to gene symbol
}
\keyword{datasets}
